[core] Improve code quality.
* [core] Allow arbitrary larger controller update period.
* [core] Fix deprecation and unused variable warnings.
* [core] Improve error reporting by printing fully qualified function name without arguments.
* [core] Remove irrelevant virtual keywords.
* [core] Use randomPermutationPeriod=0 to disable random permutation of 'PGSSolver' iterative solver, instead of 'isfinite', which is ill-defined for integers.
* [core] Do NOT shuffle first iteration cycle of 'PGSolver', and initialize indices in reverse order to make convergence (hopefully) faster.
* [core] Enable implicit conversion warnings.
* [core] More consistent coding style.
* [core] Avoid static_cast if possible and use safer conversion if possible.
* [core] Remove all C-style cast and enable warning to disallow them.
* [core] Replace (u)int32_t by std::size_t/ptrdiff_t when appropriate for clarity and portability.
* [core] Fix visco-elastic coupling forces.
* [core] Remove all dependencies to jsoncpp in core headers.
* [core/python] Add fallback to completary unsigned/signed type for integers for 'convertFromPython'.
* [gym/common/envs] Refresh action space before observation space since action could be part of observation but not conversely.
* [gym/common/wrappers] Relax argument type of 'FilteredFrameStack' for convenience.
* [gym/common/wrappers] Rename 'FilterFrameStack' in 'FilteredFrameStack' for clarity.
* [gym/examples] Automatically monitor env step 'info' on Tensorboard's histograms when using RLlib backend.
* [gym/examples] Only log data in Tensorboard format when RLlib is used, to dramatically reduce log folder size.
* [misc] Minor update of easy install dependencies.
* [misc] RLlib 'evaluate' helper method now returns agregated info over episode.
* [misc] Relax Ubuntu version check for easy install script since it should work on any release in practice.
* [misc] Do NOT install robotpkg-urdfdom binaries since 'liburdom-dev' is already installed by other easy-install dependencies.
* [misc] More robust Python library finding in provided CMake configuration file.
* [misc] Check compatibility with required jiminy version if any in provided CMake configuration file.
* [misc] Detect Numpy headers in provided CMake configuration file.
* [misc] Add missing boost definitions in CMake configuration file.

+++ core/include/jiminy/core/robot/AbstractSensor.h
@@ -94,9 +94,9 @@ namespace jiminy
 
         struct abstractSensorOptions_t
         {
-            vectorN_t const noiseStd;   ///< Standard deviation of the noise of the sensor
-            vectorN_t const bias;       ///< Bias of the sensor
-            float64_t const delay;      ///< Delay of the sensor
+            vectorN_t const noiseStd;                 ///< Standard deviation of the noise of the sensor
+            vectorN_t const bias;                     ///< Bias of the sensor
+            float64_t const delay;                    ///< Delay of the sensor
             uint32_t  const delayInterpolationOrder;  ///< Order of the interpolation used to compute delayed sensor data. [0: Zero-order holder, 1: Linear interpolation]
 
             abstractSensorOptions_t(configHolder_t const & options) :
@@ -321,7 +321,7 @@ namespace jiminy
         /// \brief      It is the size of the sensor's data vector.
         ///
         ///////////////////////////////////////////////////////////////////////////////////////////////
-        virtual uint32_t getSize(void) const = 0;
+        virtual uint64_t getSize(void) const = 0;
 
         ///////////////////////////////////////////////////////////////////////////////////////////////
         ///
@@ -450,7 +450,7 @@ namespace jiminy
         virtual int32_t const & getIdx(void) const override final;
         virtual std::string const & getType(void) const override final;
         virtual std::vector<std::string> const & getFieldnames(void) const final;
-        virtual uint32_t getSize(void) const override final;
+        virtual uint64_t getSize(void) const override final;
 
         virtual Eigen::Ref<vectorN_t const> get(void) const override final;
         virtual hresult_t setAll(float64_t const & t,
